Refactored codgen to reside in separate files
