+++ src/compiler/mod.rs
@@ -1,2 +1,7 @@
-pub(crate) mod codegen;
+pub(crate) mod class_codegen;
+pub(crate) mod control_flow_codegen;
+pub(crate) mod expr_codegen;
+pub(crate) mod ffi_codegen;
 pub(crate) mod helpers;
+pub(crate) mod main_codegen;
+pub mod primitive_codegen;
